Change license to MIT, fixes #17
Right now mozilla/nixpkgs-mozilla is MPL licensed. As this is Mozilla
code, there's not a problem with that. But nixpkgs code is traditionally
MIT and Rust code is traditionally MIT/Apache dual-licensed. If we want
this code to somehow benefit nixpkgs in the future, we'll want it
licensed under MIT, and if we want it possibly benefiting Rustup in the
future (should they decide to support Nix) we'll want to also have
Apache license.