Add support for fixed output rust channels
Using builtins.fetchurl breaks the assumption in nixpkgs that all outputs of derivations that
can access the network are fixed-output. This is the behaviour that we want when following the `latest`
rust channels, but in some cases we want full reproducability. One example is in building rust projects with hydra, hydra
by default does not accept derivations containing builtins.fetchurl.

This commit adds support for an optional attribute to the `rustChannelOf` function, the attribute
is the `sha256` of the manifest file that we wan't to use.

By calling
```nix
rustChannelOf {
  channel = "nightly";
  date = "2019-07-09";
  sha256 = "1k20hnz20var23zjhm055ymhfjpymnijfp98c387855lbsaqwhk8";
};
```
We get a fixed output derivation for the rust toolchain.

The old behaviour is kept as well, so by calling
```nix
rustChannelOf {
  channel = "nightly";
  date = "2019-07-09";
};
```

we get a non-fixed output derivation of the same compiler. This makes it so that
`latest.rustChannels.[stable|beta|nightly]` behave just as before.