Check Firefox Nightly signature after fetching the tarball.
After verification, any failure which happens during the postFecth command will
cause the failure of the derivation, and cause the output to not be reused.

This is a little painful to have these 2 different strategies for downloading
nightly and beta/releases versions, but there is no common ground to use for
both.