Fix SYSROOT detection on latest nightly
Rust changed the way the SYSROOT is determined in: https://github.com/rust-lang/rust/pull/103660

Before this change the SYSROOT was determined based on the rustc
executable. Now it is determined based on the librustc_driver-*.so file.
This pr fixes the detection by copying the library to the derivation as
we have done it for the rustc executable.